Clean up DDLX helper functions in tests
