+++ components/electric/test/electric/ddlx/ddlx_postgres_test.exs
@@ -1,34 +1,8 @@
 defmodule Electric.DDLX.DDLXPostgresTest do
   use Electric.Extension.Case, async: false
+  import ElectricTest.DDLXHelpers
 
   @moduletag ddlx: true
-
-  def query(conn, query, params \\ []) do
-    case :epgsql.equery(conn, query, params) do
-      {:ok, _n, cols, rows} ->
-        {:ok, cols, map_rows(rows)}
-
-      {:ok, cols, rows} ->
-        {:ok, cols, map_rows(rows)}
-
-      {:ok, n} when is_integer(n) ->
-        {:ok, [], []}
-
-      {:error, error} ->
-        {:error, error}
-    end
-  end
-
-  defp map_rows(rows) do
-    rows |> Enum.map(&Tuple.to_list/1) |> Enum.map(&null_to_nil/1)
-  end
-
-  defp null_to_nil(row) do
-    Enum.map(row, fn
-      :null -> nil
-      value -> value
-    end)
-  end
 
   def list_tables(conn, schema \\ "public") do
     {:ok, _cols, rows} =
@@ -78,28 +52,6 @@ defmodule Electric.DDLX.DDLXPostgresTest do
 
       {column_name, attrs}
     end
-  end
-
-  def assert_rows(conn, table_name, expected_rows) do
-    {:ok, _cols, rows} = query(conn, "select * from #{table_name}")
-
-    assert(
-      rows == expected_rows,
-      "Row assertion failed on #{table_name}, #{inspect(rows)} != #{inspect(expected_rows)}\n"
-    )
-  end
-
-  def assert_rows_slice(conn, table_name, expected_rows, range) do
-    {:ok, _cols, rows} = query(conn, "select * from #{table_name}")
-
-    rows =
-      rows
-      |> Enum.map(&Enum.slice(&1, range))
-
-    assert(
-      rows == expected_rows,
-      "Row assertion failed on #{table_name}, #{inspect(rows)} != #{inspect(expected_rows)}\n"
-    )
   end
 
   def get_foreign_keys(conn, table_name) do
