--- components/electric/test/electric/ddlx/test_helper.exs
@@ -1,1 +1,0 @@
-ExUnit.start()
